LayoutTests/svg/dynamic-updates/SVGFEFloodElement-inherit-flood-color-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the inherited 'flood-color' attribute of the SVGFEFloodElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL document.defaultView.getComputedStyle(floodElement).getPropertyValue('flood-color') should be rgb(145, 32, 103). Was rgb(0, 0, 0).
FAIL document.defaultView.getComputedStyle(floodElement).getPropertyValue('flood-color') should be rgb(64, 128, 103). Was rgb(0, 0, 0).
PASS successfullyParsed is true

TEST COMPLETE
